refactor: extract helper and connect functionality from RRS section
