+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-12T08:30:03.056Z\n"
-"PO-Revision-Date: 2019-02-12T08:30:03.056Z\n"
+"POT-Creation-Date: 2019-02-12T09:12:01.880Z\n"
+"PO-Revision-Date: 2019-02-12T09:12:01.881Z\n"
 
 msgid "Cancel"
 msgstr ""
