+++ src/pages/ReportingRateSummary.js
@@ -2,7 +2,6 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Paper } from 'material-ui'
 import { Button, DropDown } from '@dhis2/d2-ui-core'
-import { connect } from 'react-redux'
 import i18n from '../utils/i18n/locales'
 import Page from './Page'
 import DataSets from '../components/DatasetsDropdown'
@@ -12,45 +11,10 @@ import styles from '../utils/styles'
 
 import { InlineHtmlReport } from '../components/InlineHtmlReport'
 import { SectionHeadline } from '../components/SectionHeadline'
-import {
-    selectCriteria,
-    loadHtmlReport,
-} from '../redux/actions/reportingRateSummary'
-import { exportReportToXls } from '../redux/actions/htmlReport'
-
-const BASED_ON_OPTIONS = [
-    {
-        id: 'registration',
-        displayName: i18n.t('Complete data set registrations'),
-    },
-    { id: 'compulsory', displayName: i18n.t('Compulsory data elements') },
-]
-
-const isFormValid = props => props.selectedOrgUnit && props.selectedPeriod
-
-const isActionEnabled = props => isFormValid(props) && !props.loading
+import { isActionEnabled } from './reporting-rate-summary/helpers'
+import { connectReportingRateSummary } from './reporting-rate-summary/connectReportingRateSummary'
 
 export default class ReportingRateSummary extends Page {
-    static propTypes = {
-        d2: PropTypes.object.isRequired,
-        showForm: PropTypes.bool.isRequired,
-        reportHtml: PropTypes.string.isRequired,
-        selectedDataSet: PropTypes.object.isRequired,
-        selectedPeriod: PropTypes.string.isRequired,
-        selectedCriteria: PropTypes.string.isRequired,
-        loading: PropTypes.bool.isRequired,
-        exportReportToXls: PropTypes.func.isRequired,
-        loadHtmlReport: PropTypes.func.isRequired,
-        selectCriteria: PropTypes.func.isRequired,
-        selectedOrgUnit: PropTypes.object,
-    }
-
-    goToForm = () => {
-        this.setState({
-            showForm: true,
-        })
-    }
-
     render() {
         const { props } = this
         const summaryFormStyles = { display: props.showForm ? 'block' : 'none' }
@@ -60,7 +24,7 @@ export default class ReportingRateSummary extends Page {
                 <SectionHeadline
                     label={i18n.t('Resource')}
                     showBackButton={!props.showForm}
-                    onBackClick={this.goToForm}
+                    onBackClick={props.setShowForm}
                     systemVersion={props.d2.system.version}
                     sectionKey={props.sectionKey}
                 />
@@ -82,7 +46,7 @@ export default class ReportingRateSummary extends Page {
                                         fullWidth
                                         value={props.selectedCriteria}
                                         onChange={props.selectCriteria}
-                                        menuItems={BASED_ON_OPTIONS}
+                                        menuItems={props.criteriaOptions}
                                     />
                                 </div>
                                 <div id="data-set-selection">
@@ -120,28 +84,21 @@ export default class ReportingRateSummary extends Page {
     }
 }
 
-const mapStateToProps = state => ({
-    loading: state.htmlReport.loading,
-    showForm: state.reportingRateSummary.showForm,
-    reportHtml: state.htmlReport.reportHtml,
-    selectedDataSet: state.dataSet.selected,
-    selectedPeriod: state.reportPeriod.selectedPeriod,
-    selectedCriteria: state.reportingRateSummary.selectedCriteria,
-    selectedOrgUnit: state.organisationUnits,
-})
+ReportingRateSummary.propTypes = {
+    d2: PropTypes.object.isRequired,
+    showForm: PropTypes.bool.isRequired,
+    reportHtml: PropTypes.string.isRequired,
+    selectedDataSet: PropTypes.object.isRequired,
+    selectedPeriod: PropTypes.string.isRequired,
+    selectedCriteria: PropTypes.string.isRequired,
+    loading: PropTypes.bool.isRequired,
+    setShowForm: PropTypes.func.isRequired,
+    exportReportToXls: PropTypes.func.isRequired,
+    loadHtmlReport: PropTypes.func.isRequired,
+    selectCriteria: PropTypes.func.isRequired,
+    selectedOrgUnit: PropTypes.object,
+}
 
-const mapDispatchToProps = dispatch => ({
-    selectCriteria: event => dispatch(selectCriteria(event.target.value)),
-    loadHtmlReport: () => dispatch(loadHtmlReport()),
-    exportReportToXls: () =>
-        dispatch(
-            exportReportToXls(
-                document.querySelectorAll('#report-container table')
-            )
-        ),
-})
-
-export const ConnectedReportingRateSummary = connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(ReportingRateSummary)
+export const ConnectedReportingRateSummary = connectReportingRateSummary(
+    ReportingRateSummary
+)
