+++ src/pages/reporting-rate-summary/helpers.js
@@ -1,0 +1,3 @@
+export const isFormValid = props =>
+    props.selectedOrgUnit && props.selectedPeriod
+export const isActionEnabled = props => isFormValid(props) && !props.loading
